+++ lib/kitchen/element_base.rb
@@ -132,6 +132,7 @@ module Kitchen
       @ancestors = HashWithIndifferentAccess.new
       @search_query_matches_that_have_been_counted = {}
       @is_a_clone = false
+      @search_cache = {}
     end
 
     # Returns ElementBase descendent type or nil if none found
@@ -345,7 +346,7 @@ module Kitchen
     #   search results if the method or callable returns false
     # @return [ElementEnumerator]
     #
-    def search(*selector_or_xpath_args, only: nil, except: nil)
+    def search(*selector_or_xpath_args, only: nil, except: nil, reload: false)
       block_error_if(block_given?)
 
       ElementEnumerator.factory.build_within(
@@ -354,19 +355,29 @@ module Kitchen
           css_or_xpath: selector_or_xpath_args,
           only: only,
           except: except
-        )
+        ),
+        reload: reload
       )
+    end
+
+    def raw_search(*selector_or_xpath_args, reload: false)
+      key = selector_or_xpath_args
+      @search_cache[key] = nil if reload || !config.enable_search_cache
+      # cache nil search results with a fake -1 value
+      @search_cache[key] ||= raw.search(*selector_or_xpath_args) || -1
+      @search_cache[key] == -1 ? nil : @search_cache[key]
     end
 
     # Yields and returns the first child element that matches the provided
     # selector or XPath arguments.
     #
     # @param selector_or_xpath_args [Array<String>] CSS selectors or XPath arguments
+    # @param reload [Boolean] ignores cache if true
     # @yieldparam [Element] the matched XML element
     # @return [Element, nil] the matched XML element or nil if no match found
     #
     def first(*selector_or_xpath_args, reload: false)
-      cached_search(selector_or_xpath_args, method: :first, reload: reload).tap do |element|
+      search(*selector_or_xpath_args, reload: reload).first.tap do |element|
         yield(element) if block_given?
       end
     end
@@ -375,12 +386,13 @@ module Kitchen
     # selector or XPath arguments.
     #
     # @param selector_or_xpath_args [Array<String>] CSS selectors or XPath arguments
+    # @param reload [Boolean] ignores cache if true
     # @yieldparam [Element] the matched XML element
     # @raise [ElementNotFoundError] if no matching element is found
     # @return [Element] the matched XML element
     #
     def first!(*selector_or_xpath_args, reload: false)
-      cached_search(selector_or_xpath_args, method: :first!, reload: reload).tap do |element|
+      search(*selector_or_xpath_args, reload: reload).first!.tap do |element|
         yield(element) if block_given?
       end
     end
@@ -747,15 +759,6 @@ module Kitchen
     def require_one_of_child_or_sibling(child, sibling)
       raise RecipeError, 'Only one of `child` or `sibling` can be specified' if child && sibling
       raise RecipeError, 'One of `child` or `sibling` must be specified' if !child && !sibling
-    end
-
-    def cached_search(*selector_or_xpath_args, method:, reload: false)
-      key = [method, selector_or_xpath_args]
-      @search_cache ||= {}
-      @search_cache[key] = nil if reload
-      # cache nil search results with a fake -1 value
-      @search_cache[key] ||= search(*selector_or_xpath_args).send(method.to_sym) || -1
-      @search_cache[key] == -1 ? nil : @search_cache[key]
     end
 
   end