+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Caches calls to `first` and `first!` which on Chemistry halves baking time (minor)
+
 ## [4.1.0] - 2021-05-18
 
 * Fixed performance problem with element class detection (patch)